+++ CHANGELOG.md
@@ -2,13 +2,17 @@
 
 ## Unreleased 
 
+## 3.6.5
+
 ### Added 
 
 * Creator details handling & validation for manage offer and payment reviewable requests 
 
 ### Fixed
 
-* Crash core with payment transaction 
+* Crash core with payment transaction
+* Exception during invest to immediate sale
+* Wildcard value does not work for the 'payment_tasks:*'
 
 ## 3.6.4